:memo: build(version): Bump to version - 4.0.0.
This commit is for python semantic version to work.  Python Semantic
version checks the commits already in the repo for a commit that
includes the current version.  It checks the existing version  to make
sure there are no errors when calculating the next version from the new
commit messages.

close #45